쿠폰 리스트 조회 - 유저의 보유 쿠폰 없을 시 빈 리스트 반환
