+++ cubco-api/src/main/java/org/cubco/coupon/api/CouponApiDocs.java
@@ -1,16 +1,15 @@
-package org.cubco.swagger;
+package org.cubco.coupon.api;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
-import org.cubco.dto.coupon.CouponDetailResponse;
-import org.cubco.dto.coupon.CouponImageUpdateRequest;
-import org.cubco.dto.coupon.CouponResponse;
-import org.cubco.dto.coupon.CouponUseRequest;
+import org.cubco.coupon.dto.response.CouponDetailResponse;
+import org.cubco.coupon.dto.request.CouponImageUpdateRequest;
+import org.cubco.coupon.dto.response.CouponResponse;
+import org.cubco.coupon.dto.request.CouponUseRequest;
 import org.cubco.response.CommonResponse;
-import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
 
@@ -25,7 +24,6 @@ public interface CouponApiDocs {
                     content = @Content(schema = @Schema(implementation = CouponResponse.class))
             ),
             @ApiResponse(responseCode = "401", description = "인증이 필요합니다"),
-            @ApiResponse(responseCode = "404", description = "존재하는 쿠폰을 찾을 수 없습니다."),
             @ApiResponse(responseCode = "500", description = "서버 내부 오류가 발생했습니다.")
     })
     CommonResponse<List<CouponResponse>> getUserCoupons(Long userId);
