+++ cubco-api/src/main/java/org/cubco/coupon/api/CouponController.java
@@ -1,12 +1,17 @@
-package org.cubco.controller;
+package org.cubco.coupon.api;
 
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.cubco.dto.coupon.*;
+import org.cubco.auth.resolver.UserId;
+import org.cubco.coupon.dto.request.CouponImageUpdateRequest;
+import org.cubco.coupon.dto.request.CouponUseRequest;
+import org.cubco.coupon.dto.response.CouponDetailResponse;
+import org.cubco.coupon.dto.response.CouponImageUpdateResponse;
+import org.cubco.coupon.dto.response.CouponResponse;
+import org.cubco.coupon.dto.response.CouponUseResponse;
 import org.cubco.response.CommonResponse;
-import org.cubco.service.CouponService;
-import org.cubco.swagger.CouponApiDocs;
+import org.cubco.coupon.service.CouponService;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
 
@@ -22,8 +27,7 @@ public class CouponController implements CouponApiDocs {
 
     // 1. 사용자 보유 쿠폰 리스트 조회
     @GetMapping("/list")
-//    public CommonResponse<List<CouponListResponse>> getUserCoupons(@UserId Long userId) { // PR 후 주석해제
-    public CommonResponse<List<CouponResponse>> getUserCoupons(Long userId) {
+    public CommonResponse<List<CouponResponse>> getUserCoupons(@UserId Long userId) { // PR 후 주석해제
         List<CouponResponse> coupons = couponService.getCouponsByUser(userId);
         return CommonResponse.createSuccess(HttpStatus.OK, "보유 쿠폰 목록 조회 성공", coupons);
     }
@@ -31,8 +35,7 @@ public class CouponController implements CouponApiDocs {
     // 2. 쿠폰 상세 조회
     @GetMapping("/{couponId}")
     public CommonResponse<CouponDetailResponse> getCouponDetail(
-//            @UserId Long userId, // PR 후 주석해제
-            Long userId,
+            @UserId Long userId,
             @PathVariable Long couponId
     ) {
         CouponDetailResponse coupon = couponService.getCouponDetail(userId, couponId);
@@ -42,8 +45,7 @@ public class CouponController implements CouponApiDocs {
     // 3. 쿠폰 이미지 변경
     @PatchMapping("/{couponId}/image")
     public CommonResponse<?> updateCouponImage(
-//            @UserId Long userId, // PR 후 주석 해제
-            Long userId,
+            @UserId Long userId,
             @PathVariable Long couponId,
             @RequestBody @Valid CouponImageUpdateRequest request
     ) {
@@ -54,8 +56,7 @@ public class CouponController implements CouponApiDocs {
     // 4. 쿠폰 사용 (count 차감)
     @PatchMapping("/{couponId}/use")
     public CommonResponse<?> useCoupon(
-//            @UserId Long userId, // PR 후 주석 해제
-            Long userId,
+            @UserId Long userId,
             @PathVariable Long couponId,
             @RequestBody @Valid CouponUseRequest request
     ) {
