+++ cubco-api/src/main/java/org/cubco/coupon/dto/response/CouponImageUpdateResponse.java
@@ -1,4 +1,4 @@
-package org.cubco.dto.coupon;
+package org.cubco.coupon.dto.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -6,14 +6,14 @@ import org.cubco.coupon.domain.Coupon;
 
 @Getter
 @AllArgsConstructor
-public class CouponDetailResponse {
+public class CouponImageUpdateResponse {
     private Long couponId;
     private Long cafeId;
     private int stampCount;
     private String couponImageUrl;
 
-    public static CouponDetailResponse of(Coupon coupon) {
-        return new CouponDetailResponse(
+    public static CouponImageUpdateResponse of(Coupon coupon) {
+        return new CouponImageUpdateResponse(
                 coupon.getId(),
                 coupon.getCafe().getId(),
                 coupon.getCount(),
