+++ cubco-api/src/main/java/org/cubco/coupon/dto/response/CouponUseResponse.java
@@ -1,4 +1,4 @@
-package org.cubco.dto.coupon;
+package org.cubco.coupon.dto.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
